Add files via upload
The bike buyer dataset was cleaned and explored. 
I was able to explore the range at which bikes was purchased based on the dataset that was available. 
A dashboard was created which clearly shows how bike was purchase based on gender, age and other data that was available at the time the project was done.